Moving vlc handlers
